backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
polars==0.20.3
duckdb==0.9.2
supabase==2.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
minio==7.2.0
pydantic==2.5.2
pydantic-settings==2.1.0 